fixing SubtractAndReplace Object to match subtract refactoring
issue: MatterHackers/MCCentral#4764
Deleting a subtract and replace causes the original objects to move
